Feature/set client user voting

+++ client/src/pages/GamePage/ProgressSection/ProgressSection.tsx
@@ -5,11 +5,12 @@ import { deleteUserById } from 'core/api/users.service';
 import { Roles } from 'core/types/roleType';
 import { useTypeSelector } from 'core/hooks/useTypeSelector';
 import { getUsers } from 'store/actionCreators/user';
-import { AppModal } from 'core/components/modal/Modal';
+import { AppModal } from 'core/components/modals/Modal';
 import { Text } from 'core/components/Text';
 import { UserCard } from 'core/components/userCard/UserCard';
 import { ProgressCard } from 'core/components/progressCard/ProgressCard';
 import styles from './ProgressSection.module.scss';
+import { startPlayerVoting } from 'core/api/socket.service';
 
 export const ProgressSection: React.FC = () => {
   const { users } = useTypeSelector(state => state.users);
@@ -32,7 +33,16 @@ export const ProgressSection: React.FC = () => {
   };
 
   const handleSubmit = async () => {
-    await deleteUserById(getDeleteUserId);
+    if (currentUser.role === Roles.creator) {
+      deleteUserById(getDeleteUserId);
+    } else {
+      const target = {
+        gameId: currentUser.gameId,
+        playerId: currentUser.userId,
+        targetId: getDeleteUserId,
+      };
+      startPlayerVoting(target);
+    }
     setModalShowFlag(false);
     dispatch(getUsers(currentUser.gameId));
   };
@@ -77,7 +87,7 @@ export const ProgressSection: React.FC = () => {
         })}
       </Grid>
       <AppModal
-        title={`Kick player?`}
+        title={`Kick player`}
         isShow={getModalShowFlag}
         handleSubmit={handleSubmit}
         handleCancel={handleCancel}
